+++ src/index.js
@@ -1,7 +1,10 @@
 import { ValidationProvider, ValidationObserver } from "vee-validate";
 import VDynamicForm from "@/components/VDynamicForm.vue";
 
-const install = function(Vue) {
+const install = function(Vue, options = {}) {
+  Vue.prototype.dynamicFormOptions = {
+    interactionMode: options.interactionMode || "aggressive",
+  };
   Vue.component("ValidationProvider", ValidationProvider);
   Vue.component("ValidationObserver", ValidationObserver);
   Vue.component("VDynamicForm", VDynamicForm);