+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
 ## 1.1.0
+
 - feat: add component classes `v-dynamic-form`, `v-dynamic-form--inputs`, `v-dynamic-form--actions`
 - feat: add `disabled`, `readonly`, `hide-actions` props
 - feat: register **vee-validate** components globally
@@ -11,17 +12,26 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 - task: update docs
 
 ## 1.1.1
+
 - fix: register vuetify input components
 
 ## 1.2.0
+
 - fix: fix rows/cols truthy condition
 - fix: git `input-fields` prop a default value
 - fix: patch methods
 - feat: `hide-name` option per field
 
 ## 1.3.0
+
 - feat: `hidden` field option
 - feat: `component` field option for rendering custom components
 
 ## 1.3.1
+
 - feat: add `vid` field option
+
+## 1.3.2
+
+- feat: global config defaults
+- fix: apply all field options